get2gnow #0000ECaFC - ...abunch more...for later...
After fixing my last bug, my first bug started randomly accuring again.  I've finally resolved the issue by switching from float to double values for update & user ids.  I've had to store the values in gconf as a string & use strtod when I pull the values out again.  I know its a hack but its the only way to avoid issues that come up when using floats.  For what ever reason float values weren't converted from stings correctly.  It was literally coming up with a completely different value.

I've alo implement libgnome & using GnomProgram.  I've done this for many  future proofing reasons but right now my primary new feaure why I've implemented GnomeProgam right now is to have get2gnow run in 'C' locale, before threading is started, and than to ensure the orginal locale is always reset once get2gnow exits.  This is so multiple date, time, & parting methods are all now thread safe.

Lastly I'm increased the persion of get2gnow sorting and other time / date values used by get2gnow for sorting and time stamping.  Basically I've added a check so now new updates that end up calulating to less than 1 second ago, which can happen, and now when it does get2gnow's sorting value will be set to a minimum of 1 to ensure even the abs newest updates appear in the correct order.

Committed on: 2009-06-26 @ 08:19:48 PM

+++ src/tweet-view.h
@@ -79,7 +79,7 @@ GtkHBox *tweet_view_get_embed(void);
 
 void tweet_view_beep(void);
 
-void tweet_view_show_tweet(OnlineService *service, const gfloat id, const gfloat user_id, const gchar *user_name, const gchar *user_nick, const gchar *date, const gchar *sexy_tweet, const gchar *text_tweet, GdkPixbuf *pixbuf );
+void tweet_view_show_tweet(OnlineService *service, const gdouble id, const gdouble user_id, const gchar *user_name, const gchar *user_nick, const gchar *date, const gchar *sexy_tweet, const gchar *text_tweet, GdkPixbuf *pixbuf );
 
 void tweet_view_dm_data_fill(GList *followers);
